+++ src/core/esbuild-options.ts
@@ -27,6 +27,7 @@ type NormalizedBundlerOptions = {
   rootDir: string;
   sourcemap?: BundlerOptions["sourcemap"];
   splitting: boolean;
+  stripComments: boolean;
   target?: string | string[];
 };
 
@@ -49,7 +50,7 @@ function normalizeBundlerOptions(options: BundlerOptions): NormalizedBundlerOpti
     logger: options.logger,
     loggerAdapter: options.loggerAdapter,
     manifest: normalizeManifestOptions(options.manifest),
-    minify: Boolean(options.minify),
+    minify: options.minify !== false,
     onEntrySetChanged: options.onEntrySetChanged,
     onRebuilt: options.onRebuilt,
     outDir: resolvedOutDir,
@@ -58,6 +59,7 @@ function normalizeBundlerOptions(options: BundlerOptions): NormalizedBundlerOpti
     rootDir,
     sourcemap: options.sourcemap,
     splitting: Boolean(options.splitting),
+    stripComments: options.stripComments !== false,
     target: options.target,
   };
 }
@@ -78,6 +80,14 @@ function createEsbuildOptions(
     logger.info("annotate", "inline source annotations enabled");
   }
 
+  if (options.minify) {
+    logger.info("build", "minify enabled");
+  }
+
+  if (options.stripComments && !options.annotateSources) {
+    logger.info("build", "comment stripping enabled");
+  }
+
   logger.info("scss", "scss compiler enabled");
 
   return {
@@ -87,7 +97,7 @@ function createEsbuildOptions(
     entryPoints,
     external: options.external,
     format: options.format,
-    legalComments: options.annotateSources ? "inline" : undefined,
+    legalComments: options.annotateSources ? "inline" : options.stripComments ? "none" : undefined,
     logLevel: "silent",
     metafile: true,
     minify: options.minify,